Add favicon and move data.js
