+++ src/index.ejs
@@ -5,6 +5,7 @@
     <meta name="viewport" content="initial-scale=1.0, width=device-width">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>7h3 5h0p</title>
+    <link rel="shortcut icon" href="./img/icons/oggetto_logo.png" type="image/png">
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel='stylesheet'>
     <link rel="stylesheet" href="css/main.css">
     <script type="module" src="js/scripts.js" defer></script>
@@ -74,7 +75,7 @@
         <div class="product">
             <div class="product-picture">
                 <img id="productPicture" class="product-picture__img" src="img/tshirts/tshirt_white.jpg"
-                     alt=<%= nameProduct %>>
+                     alt="<%= nameProduct %>">
             </div>
             <div class="info">
                 <h1 class="info__name"><%= nameProduct %></h1>
